+++ glomar_gridding/covariance_tools.py
@@ -102,153 +102,6 @@ from statsmodels.stats import correlation_tools
 from warnings import warn
 
 
-def eof_chop(
-    cov: np.ndarray,
-    target_explained_variance: float = 0.95,
-) -> tuple[np.ndarray, dict[str, Any]]:
-    """
-    Re-compute the covariance using only eigenvectors associated with the
-    largest eigenvalues such that the explained variance achieves a target
-    value.
-
-    This method is similar to a standard eigenvalue clipping method, however
-    only the eigenvectors associated with the largest eigenvalues are computed,
-    saving on memory and improves time execution. This method is best suited to
-    larger covariance matrices, for example those for a 1-degree resolution
-    grid (approx 40_000 x 40_000). This is also appropriate for covariance
-    matrices with very large dominant modes.
-
-    This method does not preserve the total variance, i.e. the trace of the
-    output covariance matrix may not match the input.
-
-    Parameters
-    ----------
-    cov : numpy.ndarray
-        Input covariance matrix to be adjusted to positive definite.
-    target_explained_variance : float
-        Select only the largest eigenvalues such that the explained variance of
-        these eigenvalues is <= this value. The eigenvalues are first sorted in
-        descending order, then cumulatively summed. Eigenvalues that correspond
-        to values in the cumulative sum above this explained variance are
-        dropped.
-
-    Returns
-    -------
-    cov_adj : numpy.ndarray
-        Adjusted covariance matrix
-    summary_dict : dict[str, Any]
-        A dictionary containing a summary of the input and results with the
-        following keys:
-
-            - "target_explained_variance%"
-            - "num_of_retained_eofs"
-            - "threshold"
-            - "smallest_eigv"
-            - "largest_eigv"
-            - "determinant"
-            - "total_variance"
-    """
-    if np.ma.isMaskedArray(cov):
-        cov = cov.data  # type: ignore
-
-    # Compute all eigenvalues plus other useful diagonstics
-    # Eigenvalues are returned in ascending order
-    all_eigval = np.linalg.eigvalsh(cov)
-    N = len(all_eigval)
-    min_eigval = all_eigval[0]
-    max_eigval = all_eigval[-1]
-    sum_eigval = np.sum(all_eigval)
-
-    n_negatives = int(np.argmax(all_eigval >= 0))
-    sum_of_negatives = np.sum(all_eigval[:n_negatives])
-
-    p90_index = int(np.ceil(0.90 * N))
-    p95_index = int(np.ceil(0.95 * N))
-    sum_top10_eigval = np.sum(all_eigval[p90_index:])
-    sum_top05_eigval = np.sum(all_eigval[p95_index:])
-
-    top10_explained_var = 100.0 * sum_top10_eigval / sum_eigval
-    top05_explained_var = 100.0 * sum_top05_eigval / sum_eigval
-    explained_var_from_neg = 100.0 * sum_of_negatives / sum_eigval
-    print("Pre-adjusted eigenvalue summary")
-    print("Total variance=", sum_eigval)
-    print("Largest=", max_eigval)
-    print("Smallest/most negative=", min_eigval)
-    print("Number of negative eigenvalues=", n_negatives)
-    print("Sum=", sum_eigval)
-    print("Explained variance from top 10%=", top10_explained_var, "%")
-    print("Explained variance from top  5%=", top05_explained_var, "%")
-    print(
-        "Negative contributions from the negatives=",
-        explained_var_from_neg,
-        "%",
-    )
-
-    target_total_variance = target_explained_variance * sum_eigval
-    all_eigval_R = all_eigval[::-1]
-    eigenvals_2B_included = all_eigval_R[
-        all_eigval_R.cumsum() <= target_total_variance
-    ]
-    n_eig_2B_included = len(eigenvals_2B_included)
-    print("Target explained variance=", target_explained_variance)
-    print("aka adjusted total variance=", target_total_variance)
-    print("Requiring ", n_eig_2B_included, " eigenvalues")
-    print(
-        "Smallest eigenvalue in truncation=",
-        eigenvals_2B_included[-1],
-        "(aka threshold)",
-    )
-    print("Largest eigenvalue in truncation=", eigenvals_2B_included[0])
-
-    print(
-        "Computing eigenval & eigenvec up to the estimated number: "
-        + str(n_eig_2B_included)
-    )
-    subset_by_index = [N - n_eig_2B_included, N - 1]
-    # NOTE: Use SciPy here as can subset by index (numpy cannot)
-    current_eigv, current_eigV = sp.linalg.eigh(
-        cov, subset_by_index=subset_by_index
-    )
-    print(current_eigv)
-    print(current_eigV.shape)
-
-    # This is new truncated covariance matrix...
-    # it will/should have eigenvalues effectively 0
-    cov_adj = current_eigV @ np.diag(current_eigv) @ current_eigV.T
-
-    n_vec = 10
-    print("Computing adjusted eigenvalues, smallest " + str(n_vec))
-    # NOTE: Use SciPy here as can subset by index (numpy cannot)
-    new_eigv = sp.linalg.eigvalsh(cov_adj, subset_by_index=[0, n_vec - 1])
-    new_min_eigv = np.min(new_eigv)
-    new_max_eigv = np.max(new_eigv)
-    print("Largest eigenvalue=", new_max_eigv)
-    print("Smallest eigenvalue=", new_min_eigv)
-    print("Float32 precision=largest eigv x 1E-6=", 1e-6 * new_max_eigv, "]")
-
-    new_det = np.linalg.det(cov_adj)
-    print("Determinant=", new_det)
-    sum_eigval2 = np.sum(current_eigv)
-    print(
-        "Actual adjusted total variance after truncation=",
-        sum_eigval2,
-        "[Target = ",
-        target_total_variance,
-        "]",
-    )
-
-    summary_dict = {
-        "target_explained_variance%": target_explained_variance * 100.0,
-        "num_of_retained_eofs": n_eig_2B_included,
-        "threshold": eigenvals_2B_included[-1],
-        "smallest_eigv": new_min_eigv,
-        "largest_eigv": new_max_eigv,
-        "determinant": new_det,
-        "total_variance": sum_eigval2,
-    }
-    return cov_adj, summary_dict
-
-
 def check_symmetric(
     a: np.ndarray,
     rtol: float = 1e-5,
@@ -399,6 +252,7 @@ def simple_clipping(
         - https://nhigham.com/2013/02/13/the-nearest-correlation-matrix/
         - https://academic.oup.com/imajna/article/22/3/329/708688
     """
+    n = cov.shape[0]
     all_eigval = np.linalg.eigvals(cov)
     all_eigval = np.sort(all_eigval)
     max_eigval = np.max(all_eigval)
@@ -453,6 +307,8 @@ def simple_clipping(
         )
 
     n_negative = int(np.sum(all_eigval < threshold))
+    if n_negative == n:
+        warn("Input has all negative eigenvalues")
     print("Minimum eigenvalue threshold = ", threshold)
     print("Estimated number of eigenvalues below threshold = ", n_negative)
     n_vec = n_negative