+++ glomar_gridding/distances.py
@@ -30,8 +30,8 @@ def rot_mat(angle: float) -> np.ndarray:
 
     The input angle must be in radians
     """
-    c_ang = np.cos(angle)
-    s_ang = np.sin(angle)
+    c_ang = float(np.cos(angle))
+    s_ang = float(np.sin(angle))
     return np.array([[c_ang, -s_ang], [s_ang, c_ang]])
 
 
@@ -206,6 +206,7 @@ def calculate_distance_matrix(
     dist_func: Callable = haversine_distance_from_frame,
     lat_col: str = "lat",
     lon_col: str = "lon",
+    **dist_kwargs,
 ) -> np.ndarray:
     """
     Create a distance matrix from a DataFrame containing positional information,
@@ -232,6 +233,8 @@ def calculate_distance_matrix(
         Name of the column in the input DataFrame containing latitude values.
     lon_col : str
         Name of the column in the input DataFrame containing longitude values.
+    **dist_kwargs
+        Keyword arguments to pass to the distance function.
 
     Returns
     -------
@@ -239,7 +242,8 @@ def calculate_distance_matrix(
         A matrix of pairwise distances.
     """
     return dist_func(
-        df.select([pl.col(lat_col).alias("lat"), pl.col(lon_col).alias("lon")])
+        df.select([pl.col(lat_col).alias("lat"), pl.col(lon_col).alias("lon")]),
+        **dist_kwargs,
     )
 
 
@@ -451,39 +455,6 @@ def mahal_dist_func(
         delta_x * (delta_x * sigma_inv[0, 0] + delta_y * sigma_inv[0, 1])
         + delta_y * (delta_x * sigma_inv[1, 0] + delta_y * sigma_inv[1, 1])
     )
-
-
-# def _tau_unit_test():
-#     Lx = 1000.0
-#     Ly = 250.0
-#     theta = np.pi / 4
-#     sigma = Ls2sigma(Lx, Ly, theta)
-#     print(sigma)
-#     latlon0 = (10.0, -35.0)
-#     # df = pd.DataFrame({'lat': [7.0, 12.0, -1.0, 20.0],
-#     #                    'lon': [-32.0, -24.5, -27.0, -40.0]})
-#     lats = np.linspace(latlon0[0] - 10, latlon0[0] + 10, 21)
-#     lons = np.linspace(latlon0[1] - 10, latlon0[1] + 10, 21)
-#     lons2, lats2 = np.meshgrid(lons, lats)
-#     df = pd.DataFrame(
-#         {"lat": lats2.flatten().tolist(), "lon": lons2.flatten().tolist()}
-#     )
-#     print(df)
-#     pos = np.asarray(df[["lat", "lon"]].values)
-#     print(pos)
-#     pos2 = latlon2ne(pos, latlons_in_rads=False, latlon0=latlon0)
-#     print(pos2)
-#     df["northing"] = pos2[:, 0]
-#     df["easting"] = pos2[:, 1]
-#     paired_dis_mat = paired_vector_dist(pos2)
-#     print("dis_Y:")
-#     print(paired_dis_mat[:, :, 0])
-#     print("dis_X:")
-#     print(paired_dis_mat[:, :, 1])
-#     tau_mat = compute_tau_wrapper(paired_dis_mat, sigma)
-#     print("tau:")
-#     print(tau_mat)
-#     return {"tau": tau_mat, "sigma": sigma, "grid": df}
 
 
 def displacements(