+++ test/test_kriging.py
@@ -29,6 +29,7 @@ import os
 import numpy as np
 import polars as pl
 import xarray as xr
+from itertools import product
 
 from sklearn.metrics.pairwise import euclidean_distances
 
@@ -38,7 +39,15 @@ from glomar_gridding.grid import (
     grid_to_distance_matrix,
 )
 from glomar_gridding.variogram import MaternVariogram
-from glomar_gridding.kriging import kriging_ordinary, OrdinaryKriging
+from glomar_gridding.kriging import (
+    SimpleKriging,
+    constraint_mask,
+    kriging_ordinary,
+    OrdinaryKriging,
+    kriging_simple,
+    _extended_inverse,
+)
+from glomar_gridding.stochastic import StochasticKriging
 
 
 def _load_results() -> np.ndarray:
@@ -52,8 +61,10 @@ def _load_results() -> np.ndarray:
     return np.reshape(converted, (20, 20), "F")
 
 
+EXPECTED = _load_results()
+
+
 def test_ordinary_kriging() -> None:  # noqa: D103
-    expected = _load_results()
     grid = grid_from_resolution(1, [(1, 21), (1, 21)], ["lat", "lon"])
     obs = pl.DataFrame(
         {
@@ -66,7 +77,7 @@ def test_ordinary_kriging() -> None:  # noqa: D103
 
     variogram = MaternVariogram(range=35 / 3, psill=4.0, nugget=0.0, nu=1.5)
 
-    covariance: xr.DataArray = variogram.fit(dist)
+    covariance: xr.DataArray = variogram.fit(dist)  # type: ignore
 
     grid_idx = obs.get_column("grid_idx").to_numpy()
     obs_vals = obs.get_column("val").to_numpy()
@@ -75,12 +86,11 @@ def test_ordinary_kriging() -> None:  # noqa: D103
     SS = covariance.values[grid_idx, :]
     k, _ = kriging_ordinary(S, SS, obs_vals, covariance.values)
 
-    assert np.allclose(expected, np.reshape(k, (20, 20), "C"))  # noqa: S101
+    assert np.allclose(EXPECTED, np.reshape(k, (20, 20), "C"))  # noqa: S101
     return None
 
 
 def test_ordinary_kriging_class() -> None:  # noqa: D103
-    expected = _load_results()
     grid = grid_from_resolution(1, [(1, 21), (1, 21)], ["lat", "lon"])
     obs = pl.DataFrame(
         {
@@ -93,7 +103,7 @@ def test_ordinary_kriging_class() -> None:  # noqa: D103
 
     variogram = MaternVariogram(range=35 / 3, psill=4.0, nugget=0.0, nu=1.5)
 
-    covariance: xr.DataArray = variogram.fit(dist)
+    covariance: xr.DataArray = variogram.fit(dist)  # type: ignore
 
     grid_idx = obs.get_column("grid_idx").to_numpy()
     obs_vals = obs.get_column("val").to_numpy()
@@ -102,12 +112,11 @@ def test_ordinary_kriging_class() -> None:  # noqa: D103
 
     k = OKrige.solve(obs_vals, grid_idx)
 
-    assert np.allclose(expected, np.reshape(k, (20, 20), "C"))  # noqa: S101
+    assert np.allclose(EXPECTED, np.reshape(k, (20, 20), "C"))  # noqa: S101
     return None
 
 
 def test_ordinary_kriging_class_from_weights() -> None:  # noqa: D103
-    expected = _load_results()
     grid = grid_from_resolution(1, [(1, 21), (1, 21)], ["lat", "lon"])
     obs = pl.DataFrame(
         {
@@ -120,7 +129,7 @@ def test_ordinary_kriging_class_from_weights() -> None:  # noqa: D103
 
     variogram = MaternVariogram(range=35 / 3, psill=4.0, nugget=0.0, nu=1.5)
 
-    covariance: xr.DataArray = variogram.fit(dist)
+    covariance: xr.DataArray = variogram.fit(dist)  # type: ignore
 
     grid_idx = obs.get_column("grid_idx").to_numpy()
     obs_vals = obs.get_column("val").to_numpy()
@@ -139,12 +148,11 @@ def test_ordinary_kriging_class_from_weights() -> None:  # noqa: D103
 
     k = OKrige.solve(obs_vals, grid_idx)
 
-    assert np.allclose(expected, np.reshape(k, (20, 20), "C"))  # noqa: S101
+    assert np.allclose(EXPECTED, np.reshape(k, (20, 20), "C"))  # noqa: S101
     return None
 
 
 def test_ordinary_kriging_class_from_inv() -> None:  # noqa: D103
-    expected = _load_results()
     grid = grid_from_resolution(1, [(1, 21), (1, 21)], ["lat", "lon"])
     obs = pl.DataFrame(
         {
@@ -157,7 +165,7 @@ def test_ordinary_kriging_class_from_inv() -> None:  # noqa: D103
 
     variogram = MaternVariogram(range=35 / 3, psill=4.0, nugget=0.0, nu=1.5)
 
-    covariance: xr.DataArray = variogram.fit(dist)
+    covariance: xr.DataArray = variogram.fit(dist)  # type: ignore
 
     grid_idx = obs.get_column("grid_idx").to_numpy()
     obs_vals = obs.get_column("val").to_numpy()
@@ -172,5 +180,170 @@ def test_ordinary_kriging_class_from_inv() -> None:  # noqa: D103
 
     k = OKrige.solve(obs_vals, grid_idx)
 
-    assert np.allclose(expected, np.reshape(k, (20, 20), "C"))  # noqa: S101
+    assert np.allclose(EXPECTED, np.reshape(k, (20, 20), "C"))  # noqa: S101
+    return None
+
+
+def test_ordinary_kriging_class_methods() -> None:  # noqa: D103
+    grid = grid_from_resolution(1, [(1, 21), (1, 21)], ["lat", "lon"])
+    obs = pl.DataFrame(
+        {
+            "lat": [5.0, 15.0, 10.0],
+            "lon": [5.0, 10.0, 15.0],
+            "val": [1.0, 0.0, 1.0],
+        }
+    ).pipe(map_to_grid, grid, grid_coords=["lat", "lon"])
+
+    grid_idx = obs.get_column("grid_idx").to_numpy()
+    obs_vals = obs.get_column("val").to_numpy()
+
+    dist: xr.DataArray = grid_to_distance_matrix(grid, euclidean_distances)
+
+    variogram = MaternVariogram(range=35 / 3, psill=4.0, nugget=0.0, nu=1.5)
+
+    covariance: xr.DataArray = variogram.fit(dist)  # type: ignore
+
+    err_cov = np.full(covariance.shape, np.nan)
+    err_cov_vals = np.random.rand(3, 3)
+    err_cov_vals = np.dot(err_cov_vals, err_cov_vals.T)
+    idx = list(product(grid_idx, grid_idx))
+    for i, val in zip(idx, err_cov_vals.flatten()):
+        err_cov[*i] = val
+
+    OKrige = OrdinaryKriging(covariance=covariance.values)
+    k = OKrige.solve(obs_vals, grid_idx, error_cov=err_cov)
+    u = OKrige.get_uncertainty(grid_idx)
+    a = OKrige.constraint_mask(grid_idx)
+
+    assert k.shape == a.shape == u.shape
+
+    S = covariance.values[grid_idx[:, None], grid_idx[None, :]] + err_cov_vals
+    SS = covariance.values[grid_idx, :]
+    k2, u2 = kriging_ordinary(S, SS, obs_vals, covariance.values)
+
+    assert np.allclose(k2, k)
+
+    assert np.allclose(u2, u)
+
+    return None
+
+
+def test_simple_kriging_class_methods() -> None:  # noqa: D103
+    grid = grid_from_resolution(1, [(1, 21), (1, 21)], ["lat", "lon"])
+    obs = pl.DataFrame(
+        {
+            "lat": [5.0, 15.0, 10.0],
+            "lon": [5.0, 10.0, 15.0],
+            "val": [1.0, 0.0, 1.0],
+        }
+    ).pipe(map_to_grid, grid, grid_coords=["lat", "lon"])
+
+    grid_idx = obs.get_column("grid_idx").to_numpy()
+    obs_vals = obs.get_column("val").to_numpy()
+
+    dist: xr.DataArray = grid_to_distance_matrix(grid, euclidean_distances)
+
+    variogram = MaternVariogram(range=35 / 3, psill=4.0, nugget=0.0, nu=1.5)
+
+    covariance: xr.DataArray = variogram.fit(dist)  # type: ignore
+
+    err_cov = np.full(covariance.shape, np.nan)
+    err_cov_vals = np.random.rand(3, 3)
+    err_cov_vals = np.dot(err_cov_vals, err_cov_vals.T)
+    idx = list(product(grid_idx, grid_idx))
+    for i, val in zip(idx, err_cov_vals.flatten()):
+        err_cov[*i] = val
+
+    SKrige = SimpleKriging(covariance=covariance.values)
+    k = SKrige.solve(obs_vals, grid_idx, error_cov=err_cov)
+    u = SKrige.get_uncertainty(grid_idx)
+    a = SKrige.constraint_mask(grid_idx)
+
+    assert k.shape == a.shape == u.shape
+
+    S = covariance.values[grid_idx[:, None], grid_idx[None, :]] + err_cov_vals
+    SS = covariance.values[grid_idx, :]
+    k2, u2 = kriging_simple(S, SS, obs_vals, covariance.values)
+
+    a2 = constraint_mask(S, SS, covariance.values)
+
+    assert np.allclose(k2, k)
+
+    assert np.allclose(u2, u)
+
+    assert np.allclose(a2, a)
+
+    return None
+
+
+@pytest.mark.parametrize(
+    "name, n",
+    [
+        ("n = 10", 10),
+        ("n = 25", 25),
+        ("n = 100", 100),
+        ("n = 1000", 1000),
+    ],
+)
+def test_inverse_trick(name, n):
+    np.random.seed(31900)
+    A = np.random.rand(n, n)
+    S = np.dot(A, A.T)
+
+    Sinv = np.linalg.inv(S)
+    Sinv_ext = _extended_inverse(Sinv)
+
+    S_ext = np.block([[S, np.ones((n, 1))], [np.ones((1, n)), 0]])
+
+    assert np.allclose(Sinv_ext, np.linalg.inv(S_ext))
+
+
+def test_stochastic_kriging_class_methods() -> None:  # noqa: D103
+    grid = grid_from_resolution(1, [(1, 21), (1, 21)], ["lat", "lon"])
+    obs = pl.DataFrame(
+        {
+            "lat": [5.0, 15.0, 10.0],
+            "lon": [5.0, 10.0, 15.0],
+            "val": [1.0, 0.0, 1.0],
+        }
+    ).pipe(map_to_grid, grid, grid_coords=["lat", "lon"])
+
+    grid_idx = obs.get_column("grid_idx").to_numpy()
+    obs_vals = obs.get_column("val").to_numpy()
+
+    dist: xr.DataArray = grid_to_distance_matrix(grid, euclidean_distances)
+
+    variogram = MaternVariogram(range=35 / 3, psill=4.0, nugget=0.0, nu=1.5)
+
+    covariance: xr.DataArray = variogram.fit(dist)  # type: ignore
+
+    err_cov = np.full(covariance.shape, np.nan)
+    err_cov_vals = np.random.rand(3, 3)
+    err_cov_vals = np.dot(err_cov_vals, err_cov_vals.T)
+    idx = list(product(grid_idx, grid_idx))
+    for i, val in zip(idx, err_cov_vals.flatten()):
+        err_cov[*i] = val
+
+    StochKrige = StochasticKriging(covariance=covariance.values)
+    k = StochKrige.solve(obs_vals, grid_idx, error_cov=err_cov)
+    u = StochKrige.get_uncertainty(grid_idx)
+    a = StochKrige.constraint_mask(grid_idx)
+
+    assert k.shape == a.shape == u.shape
+
+    S = covariance.values[grid_idx[:, None], grid_idx[None, :]] + err_cov_vals
+    SS = covariance.values[grid_idx, :]
+    k2, u2 = kriging_ordinary(S, SS, obs_vals, covariance.values)
+
+    assert hasattr(StochKrige, "gridded_field")
+    assert np.allclose(k2, StochKrige.gridded_field)
+    assert np.allclose(u2, u)
+
+    sk_weights = StochKrige.simple_kriging_weights
+    delattr(StochKrige, "simple_kriging_weights")
+    StochKrige.set_simple_kriging_weights(sk_weights)
+    a2 = StochKrige.constraint_mask(grid_idx)
+
+    assert np.allclose(a, a2)
+
     return None